contributors/adobe/submitted/shapes/wrap-padding-001.html

<!DOCTYPE HTML>
<html>
    <head>
        <title>CSS Exclusions Test: wrap-padding property set</title>
        <link rel="author" title="Jacob Goldstein" href="mailto:jacobg@adobe.com">
        <link rel="help" href="http://www.w3.org/TR/css-shapes-1/#shape-margin-property">
        <meta name="flags" content="ahem dom">
        <meta name="assert" content="Text should flow within the exclusion, offset by the wrap-padding">
        <script src="/resources/testharness.js"></script>
        <script src="/resources/testharnessreport.js"></script>
        <script src="resources/helper.js"></script>
        <style>
            * {
                margin: 0px;
                padding: 0px;
            }
            .exclusion {
                wrap-flow: clear;
                wrap-padding: 12px;
                position: absolute;
                top: 40px;
                left: 50px;
                width: 135px;
                height: 110px;
                background: blue;
                font: 12px Ahem;
            }    
            #content {
                width: 400px;
                height: 200px;
                line-height: 12px;
                font: 12px Ahem;
            }
        </style>
    </head>
    <body>
        <div class="exclusion"><span id="insideDiv_w1">Hello</span><span id="insideDiv_w2">World</span></div>

        <div id="content"><span id="linesAbove">A<BR>B<BR>C<BR></span><span id="linesBelow">D<BR>E</span>
        </div>
        
        <div id="log"></div>
        
        <script type="text/javascript">            
            test(function() {assert_equals(checkLinePos("insideDiv_w1",62,"left"), true)}, "Verify left of 'insideDiv_w1' span appears offset based on wrap-padding");
            test(function() {assert_equals(checkLinePos("insideDiv_w2",64,"top"), true)}, "Verify top of 'insideDiv_w2' span appears offset based on wrap-padding");
            test(function() {assert_equals(checkLinePos("insideDiv_w2",62,"left"), true)}, "Verify left of 'insideDiv_w2' span appears offset based on wrap-padding");
            test(function() {assert_equals(checkLinePos("linesBelow",150,"top"), true)}, "Verify top of 'linesBelow' span appears below the exclusion");
        </script>
    </body>
</html>
